pyproject.toml
[tool.poetry]
name = "azure-functions-python-poetry"
version = "0.1.0"
description = ""
authors = ["Rasmus Hummelmose <rasmus@hummelmose.dk>"]

[tool.poetry.dependencies]
python = "^3.6"
azure-functions = "1.0.0a5"
azure-storage-queue = "^1.4"

[tool.poetry.dev-dependencies]
pylint = "^2.0"
azure-functions-worker = "1.0.0a6"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
